requirements.txt
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
streamlit==1.29.0
